finalizing managed group testing framework
