+++ tests/test_managed_group.rs
@@ -233,8 +233,8 @@ fn managed_group_operations() {
 
             // Check that Alice & Bob are the members of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Bob");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Bob");
 
             let mut bob_group = ManagedGroup::new_from_welcome(
                 &bob_credential_bundle,
@@ -362,9 +362,9 @@ fn managed_group_operations() {
 
             // Check that Alice, Bob & Charlie are the members of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Bob");
-            assert_eq!(members[2].1.identity(), b"Charlie");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Bob");
+            assert_eq!(members[2].identity(), b"Charlie");
 
             // === Charlie sends a message to the group ===
             let message_charlie = b"Hi, I'm Charlie!";
@@ -452,8 +452,8 @@ fn managed_group_operations() {
 
             // Check that Alice & Charlie are the members of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Charlie");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Charlie");
 
             // Check that Bob can no longer send messages
             assert!(bob_group.create_message(b"Should not go through").is_err());
@@ -504,8 +504,8 @@ fn managed_group_operations() {
 
             // Check that Alice & Bob are the members of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Bob");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Bob");
 
             // Bob creates a new group
             let mut bob_group = ManagedGroup::new_from_welcome(
@@ -522,16 +522,16 @@ fn managed_group_operations() {
 
             // Check that Alice & Bob are the members of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Bob");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Bob");
 
             // Make sure the group contains two members
             assert_eq!(bob_group.members().len(), 2);
 
             // Check that Alice & Bob are the members of the group
             let members = bob_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
-            assert_eq!(members[1].1.identity(), b"Bob");
+            assert_eq!(members[0].identity(), b"Alice");
+            assert_eq!(members[1].identity(), b"Bob");
 
             // === lice sends a message to the group ===
             let message_alice = b"Hi, I'm Alice!";
@@ -583,7 +583,7 @@ fn managed_group_operations() {
 
             // Check that Alice is the only member of the group
             let members = alice_group.members();
-            assert_eq!(members[0].1.identity(), b"Alice");
+            assert_eq!(members[0].identity(), b"Alice");
 
             // === Auto-save ===
